+++ src/store/selectors/auth_selector/AuthSelector.tsx
