+++ src/store/selectors/auth_selector/index.ts
