+++ src/store/selectors/image_modal_selector/ImageModalSelector.tsx
