+++ src/store/selectors/image_modal_selector/index.ts
