+++ src/store/selectors/index.ts
@@ -1,5 +1,5 @@
-export * from './auth';
-export * from './image_modal';
-export * from './notification';
-// export * from './alert';
-// export * from './profile';
+export * from './auth_selector';
+export * from './image_modal_selector';
+export * from './notification_selector';
+// export * from './alert_selector';
+// export * from './profile_selector';
